Mappings

+++ 00_RestAspNet_Scaffold/RestAspNet/Controllers/PersonController.cs
@@ -26,7 +26,7 @@ namespace RestAspNet.Controllers
         }
 
         [HttpGet("{id}")]
-        public IActionResult Get(int id)
+        public IActionResult Get(long id)
         {
             var person = _personService.FindByID(id);
 
@@ -52,7 +52,7 @@ namespace RestAspNet.Controllers
         }
 
         [HttpDelete("{id}")]
-        public IActionResult Delete(int id)
+        public IActionResult Delete(long id)
         {
             _personService.Delete(id);
