+++ 00_RestAspNet_Scaffold/RestAspNet/Services/Implementations/IPersonService.cs
@@ -7,9 +7,9 @@ namespace RestAspNet.Services.Implementations
     public interface IPersonService
     {
         Person Create(Person person);
-        Person FindByID(int id);
+        Person FindByID(long id);
         List<Person> FindAll();
         Person Update(Person person);
-        void Delete(int id);
+        void Delete(long id);
     }
 }
